requirements.txt
Package            Version
------------------ ---------
appdirs            1.4.4
backcall           0.2.0
certifi            2021.5.30
charset-normalizer 2.0.6
colorama           0.4.4
cycler             0.10.0
decorator          5.1.0
fire               0.4.0
future             0.18.2
idna               3.2
ipython            7.28.0
jedi               0.18.0
kiwisolver         1.3.2
matplotlib         3.4.3
matplotlib-inline  0.1.3
monty              2021.8.17
mpmath             1.2.1
networkx           2.6.3
numpy              1.21.2
palettable         3.3.0
pandas             1.3.3
parso              0.8.2
pathspec           0.9.0
pickleshare        0.7.5
Pillow             8.3.2
pip                21.0.1
pipdate            0.3.1
plotly             5.3.1
prompt-toolkit     3.0.20
Pygments           2.10.0
pymatgen           2022.0.14
pyparsing          2.4.7
python-dateutil    2.8.2
pytz               2021.1
PyYAML             5.4.1
requests           2.26.0
ruamel.yaml        0.17.16
ruamel.yaml.clib   0.2.6
scipy              1.7.1
setuptools         58.0.4
six                1.16.0
spglib             1.16.1
structure          0.1.dev16
sympy              1.8
tabulate           0.8.9
tenacity           8.0.1
termcolor          1.1.0
tinynetrc          1.3.1
traitlets          5.1.0
uncertainties      3.1.6
urllib3            1.26.7
wcwidth            0.2.5
wheel              0.37.0
wincertstore       0.2
